Added performance tester screen
Can be used to optimize the reconciler to be as fast as possible compared to pure JS rendering using a SkiaView.

Not sure if this screen will be kept - but it is super convinient for testing. In pure JS / SkiaView mode it renders over 2000 colored rects in 6ms spent in the JS thread and 2ms spent in the GPU/render thread.