Moved updating and clearing of property changed to before / after render
To avoid updating a property in the middle of rendering we now update all props for the whole dom tree before we render, and also wait until we're done rendering with clearing any flags on props and drawing contexts.

This fixes the issue with nodes being out of sync when receiving touch events.