Removed extra use of ref counting when swapping property values
The method updatePendingChanges() for props swaps buffers, but the swap started showing up in the profiler for big DOM trees.

This commit simplifies this by using the std::shared_ptr `swap` method directly. This causes no inc/dec of the ref counts - just changes the pointers in the fastest way.