Added locking to mapped properties to fix threading issue with Reanimated
Reanimated calls setProp on the main thread (worklet) while other props are set on the Javascript thread.

The crash we saw was caused by a race condition where setProps was called from the JS thread clearing the mapped properties vector while in the middle of setProp called from the worklet thread.

This commit fixes this by adding a lock and methods for enumerating properties in a safe way.